requirements.txt
flask
ultralytics
opencv-python
numpy
timm
torch
pillow